Add regex framework so we can handle multiple different regular expression formats. Move all the PMIx internal regex support to the "native" component, and make it always be the highest priority as it is the default. When parsing a regex, each module will check for the starting string to indicate the regex generator that was used. "pmix" will indicate native.
We probably won't be generating other formats, so allow any given API to be NULL.
