+++ src/client/pmix_client.c
@@ -72,6 +72,7 @@ static const char pmix_version_string[] = PMIX_VERSION;
 #include "src/threads/threads.h"
 #include "src/mca/bfrops/base/base.h"
 #include "src/mca/gds/base/base.h"
+#include "src/mca/preg/preg.h"
 #include "src/mca/ptl/base/base.h"
 #include "src/include/pmix_globals.h"
 
@@ -1110,96 +1111,11 @@ static void _commitfn(int sd, short args, void *cbdata)
     return rc;
 }
 
-static void _resolve_peers(int sd, short args, void *cbdata)
-{
-    pmix_cb_t *cb = (pmix_cb_t*)cbdata;
-    pmix_status_t rc;
-    pmix_kval_t *kv;
-    pmix_proc_t proc;
-    char **ptr;
-    pmix_info_t *info;
-    pmix_proc_t *procs;
-    size_t ninfo, nprocs, n, j;
-
-    /* this data isn't going anywhere, so we don't require a copy */
-    cb->copy = false;
-    /* scope is irrelevant as the info we seek must be local */
-    cb->scope = PMIX_SCOPE_UNDEF;
-    /* let the proc point to the nspace */
-    (void)strncpy(proc.nspace, cb->pname.nspace, PMIX_MAX_NSLEN);
-    proc.rank = PMIX_RANK_WILDCARD;
-    cb->proc = &proc;
-
-    PMIX_GDS_FETCH_KV(rc, pmix_client_globals.myserver, cb);
-    if (PMIX_SUCCESS != rc) {
-        if (PMIX_ERR_INVALID_NAMESPACE != rc) {
-            PMIX_ERROR_LOG(rc);
-        }
-        goto complete;
-    }
-    /* should just be the one value on the list */
-    if (1 != pmix_list_get_size(&cb->kvs)) {
-        PMIX_ERROR_LOG(PMIX_ERR_BAD_PARAM);
-        rc = PMIX_ERR_BAD_PARAM;
-        goto complete;
-    }
-    kv = (pmix_kval_t*)pmix_list_get_first(&cb->kvs);
-    /* the hostname used as a key with wildcard rank will return
-     * a pmix_data_array_t of pmix_info_t structs */
-    if (NULL == kv->value ||
-        PMIX_DATA_ARRAY != kv->value->type ||
-        NULL == kv->value->data.darray ||
-        PMIX_INFO != kv->value->data.darray->type) {
-        PMIX_ERROR_LOG(PMIX_ERR_DATA_VALUE_NOT_FOUND);
-        rc = PMIX_ERR_DATA_VALUE_NOT_FOUND;
-        goto complete;
-    }
-    info = (pmix_info_t*)kv->value->data.darray->array;
-    ninfo = kv->value->data.darray->size;
-    /* find the PMIX_LOCAL_PEERS key */
-    for (n=0; n < ninfo; n++) {
-        if (0 == strncmp(info[n].key, PMIX_LOCAL_PEERS, PMIX_MAX_KEYLEN)) {
-            /* split the string */
-            ptr = pmix_argv_split(info[n].value.data.string, ',');
-            nprocs = pmix_argv_count(ptr);
-            PMIX_PROC_CREATE(procs, nprocs);
-            if (NULL == procs) {
-                rc = PMIX_ERR_NOMEM;
-                pmix_argv_free(ptr);
-                goto complete;
-            }
-            for (j=0; j < nprocs; j++) {
-                (void)strncpy(procs[j].nspace, cb->pname.nspace, PMIX_MAX_NSLEN);
-                procs[j].rank = strtoul(ptr[j], NULL, 10);
-            }
-            cb->procs = procs;
-            cb->nprocs = nprocs;
-            rc = PMIX_SUCCESS;
-            pmix_argv_free(ptr);
-            goto complete;
-        }
-    }
-
-  complete:
-    cb->status = rc;
-    if (NULL != cb->info) {
-        PMIX_INFO_FREE(cb->info, cb->ninfo);
-    }
-    cb->pstatus = rc;
-    /* post the data so the receiving thread can acquire it */
-    PMIX_POST_OBJECT(cb);
-    PMIX_WAKEUP_THREAD(&cb->lock);
-    return;
-}
 /* need to thread-shift this request */
 PMIX_EXPORT pmix_status_t PMIx_Resolve_peers(const char *nodename,
                                              const char *nspace,
                                              pmix_proc_t **procs, size_t *nprocs)
 {
-    pmix_cb_t *cb;
-    pmix_status_t rc;
-    pmix_proc_t proc;
-
     PMIX_ACQUIRE_THREAD(&pmix_global_lock);
     if (pmix_globals.init_cntr <= 0) {
         PMIX_RELEASE_THREAD(&pmix_global_lock);
@@ -1211,113 +1127,12 @@ PMIX_EXPORT pmix_status_t PMIx_Resolve_peers(const char *nodename,
     *procs = NULL;
     *nprocs = 0;
 
-    cb = PMIX_NEW(pmix_cb_t);
-    cb->key = (char*)nodename;
-    cb->pname.nspace = strdup(nspace);
-
-    PMIX_THREADSHIFT(cb, _resolve_peers);
-
-    /* wait for the result */
-    PMIX_WAIT_THREAD(&cb->lock);
-
-    /* if the nspace wasn't found, then we need to
-     * ask the server for that info */
-    if (PMIX_ERR_INVALID_NAMESPACE == cb->status) {
-        (void)strncpy(proc.nspace, nspace, PMIX_MAX_NSLEN);
-        proc.rank = PMIX_RANK_WILDCARD;
-        /* any key will suffice as it will bring down
-         * the entire data blob */
-        rc = PMIx_Get(&proc, PMIX_UNIV_SIZE, NULL, 0, NULL);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_RELEASE(cb);
-            return rc;
-        }
-        /* retry the fetch */
-        cb->lock.active = true;
-        PMIX_THREADSHIFT(cb, _resolve_peers);
-        PMIX_WAIT_THREAD(&cb->lock);
-    }
-    *procs = cb->procs;
-    *nprocs = cb->nprocs;
-
-    rc = cb->status;
-    PMIX_RELEASE(cb);
-    return rc;
-}
-
-static void _resolve_nodes(int sd, short args, void *cbdata)
-{
-    pmix_cb_t *cb = (pmix_cb_t*)cbdata;
-    pmix_status_t rc;
-    pmix_kval_t *kv;
-    pmix_proc_t proc;
-
-      /* create a pmix_info_t so we can pass the nspace
-       * into the fetch as a qualifier */
-      PMIX_INFO_CREATE(cb->info, 1);
-      if (NULL == cb->info) {
-          cb->status = PMIX_ERR_NOMEM;
-          PMIX_POST_OBJECT(cb);
-          PMIX_WAKEUP_THREAD(&cb->lock);
-          return;
-      }
-      cb->ninfo = 1;
-      PMIX_INFO_LOAD(&cb->info[0], PMIX_NSPACE, cb->pname.nspace, PMIX_STRING);
-      /* tell the GDS what we want */
-      cb->key = PMIX_NODE_LIST;
-      /* this data isn't going anywhere, so we don't require a copy */
-      cb->copy = false;
-      /* scope is irrelevant as the info we seek must be local */
-      cb->scope = PMIX_SCOPE_UNDEF;
-      /* put the nspace in the proc field */
-      (void)strncpy(proc.nspace, cb->pname.nspace, PMIX_MAX_NSLEN);
-      /* the info will be associated with PMIX_RANK_WILDCARD */
-      proc.rank = PMIX_RANK_WILDCARD;
-      cb->proc = &proc;
-
-      PMIX_GDS_FETCH_KV(rc, pmix_client_globals.myserver, cb);
-      if (PMIX_SUCCESS != rc) {
-          PMIX_ERROR_LOG(rc);
-          goto complete;
-      }
-      /* should just be the one value on the list */
-      if (1 != pmix_list_get_size(&cb->kvs)) {
-          PMIX_ERROR_LOG(PMIX_ERR_BAD_PARAM);
-          rc = PMIX_ERR_BAD_PARAM;
-          goto complete;
-      }
-      kv = (pmix_kval_t*)pmix_list_get_first(&cb->kvs);
-      /* the PMIX_NODE_LIST key is supposed to return a comma-delimited
-       * string of nodes in this  - check that it did */
-      if (NULL == kv->value ||
-          PMIX_STRING != kv->value->type) {
-          PMIX_ERROR_LOG(PMIX_ERR_DATA_VALUE_NOT_FOUND);
-          rc = PMIX_ERR_DATA_VALUE_NOT_FOUND;
-          goto complete;
-      }
-      /* return the string */
-      if (NULL != kv->value->data.string) {
-        cb->key = strdup(kv->value->data.string);
-      }
-
-    complete:
-      cb->status = rc;
-      if (NULL != cb->info) {
-          PMIX_INFO_FREE(cb->info, cb->ninfo);
-      }
-      /* post the data so the receiving thread can acquire it */
-      PMIX_POST_OBJECT(cb);
-      PMIX_WAKEUP_THREAD(&cb->lock);
-      return;
+    return pmix_preg.resolve_peers(nodename, nspace, procs, nprocs);
 }
 
 /* need to thread-shift this request */
 PMIX_EXPORT pmix_status_t PMIx_Resolve_nodes(const char *nspace, char **nodelist)
 {
-    pmix_cb_t *cb;
-    pmix_status_t rc;
-    pmix_proc_t proc;
-
     PMIX_ACQUIRE_THREAD(&pmix_global_lock);
     if (pmix_globals.init_cntr <= 0) {
         PMIX_RELEASE_THREAD(&pmix_global_lock);
@@ -1325,35 +1140,8 @@ PMIX_EXPORT pmix_status_t PMIx_Resolve_nodes(const char *nspace, char **nodelist
     }
     PMIX_RELEASE_THREAD(&pmix_global_lock);
 
-    cb = PMIX_NEW(pmix_cb_t);
-    cb->pname.nspace = (char*)nspace;
+    /* set default */
+    *nodelist = NULL;
 
-    PMIX_THREADSHIFT(cb, _resolve_nodes);
-
-    /* wait for the result */
-    PMIX_WAIT_THREAD(&cb->lock);
-
-    /* if the nspace wasn't found, then we need to
-     * ask the server for that info */
-    if (PMIX_ERR_INVALID_NAMESPACE == cb->status) {
-        (void)strncpy(proc.nspace, nspace, PMIX_MAX_NSLEN);
-        proc.rank = PMIX_RANK_WILDCARD;
-        /* any key will suffice as it will bring down
-         * the entire data blob */
-        rc = PMIx_Get(&proc, PMIX_UNIV_SIZE, NULL, 0, NULL);
-        if (PMIX_SUCCESS != rc) {
-            PMIX_RELEASE(cb);
-            return rc;
-        }
-        /* retry the fetch */
-        cb->lock.active = true;
-        PMIX_THREADSHIFT(cb, _resolve_nodes);
-        PMIX_WAIT_THREAD(&cb->lock);
-    }
-    /* the string we want is in the key field */
-    *nodelist = cb->key;
-
-    rc = cb->status;
-    PMIX_RELEASE(cb);
-    return rc;
+    return pmix_preg.resolve_nodes(nspace, nodelist);
 }
