+++ src/runtime/pmix_init.c
@@ -47,6 +47,7 @@
 #include "src/mca/pinstalldirs/base/base.h"
 #include "src/mca/pnet/base/base.h"
 #include "src/mca/psec/base/base.h"
+#include "src/mca/preg/base/base.h"
 #include "src/mca/ptl/base/base.h"
 
 #include "src/event/pmix_event.h"
@@ -256,6 +257,16 @@ int pmix_rte_init(pmix_proc_type_t type,
     }
     if (PMIX_SUCCESS != (ret = pmix_pnet_base_select())) {
         error = "pmix_pnet_base_select";
+        goto return_error;
+    }
+
+    /* open the preg and select the active plugins */
+    if( PMIX_SUCCESS != (ret = pmix_mca_base_framework_open(&pmix_preg_base_framework, 0)) ) {
+        error = "pmix_preg_base_open";
+        goto return_error;
+    }
+    if( PMIX_SUCCESS != (ret = pmix_preg_base_select()) ) {
+        error = "pmix_preg_base_select";
         goto return_error;
     }
