+++ src/server/pmix_server.c
@@ -64,6 +64,7 @@
 #include "src/runtime/pmix_rte.h"
 #include "src/mca/bfrops/base/base.h"
 #include "src/mca/gds/base/base.h"
+#include "src/mca/preg/preg.h"
 #include "src/mca/ptl/base/base.h"
 
 /* the server also needs access to client operations
@@ -1207,21 +1208,8 @@ PMIX_EXPORT pmix_status_t PMIx_Store_internal(const pmix_proc_t *proc,
     return rc;
 }
 
-#define PMIX_MAX_NODE_PREFIX        50
-
 PMIX_EXPORT pmix_status_t PMIx_generate_regex(const char *input, char **regexp)
 {
-    char *vptr, *vsave;
-    char prefix[PMIX_MAX_NODE_PREFIX];
-    int i, j, len, startnum, vnum, numdigits;
-    bool found, fullval;
-    char *suffix, *sfx;
-    pmix_regex_value_t *vreg;
-    pmix_regex_range_t *range;
-    pmix_list_t vids;
-    char **regexargs = NULL, *tmp, *tmp2;
-    char *cptr;
-
     PMIX_ACQUIRE_THREAD(&pmix_global_lock);
     if (pmix_globals.init_cntr <= 0) {
         PMIX_RELEASE_THREAD(&pmix_global_lock);
@@ -1229,226 +1217,11 @@ PMIX_EXPORT pmix_status_t PMIx_generate_regex(const char *input, char **regexp)
     }
     PMIX_RELEASE_THREAD(&pmix_global_lock);
 
-    /* define the default */
-    *regexp = NULL;
-
-    /* setup the list of results */
-    PMIX_CONSTRUCT(&vids, pmix_list_t);
-
-    /* cycle thru the array of input values - first copy
-     * it so we don't overwrite what we were given*/
-    vsave = strdup(input);
-    vptr = vsave;
-    while (NULL != (cptr = strchr(vptr, ',')) || 0 < strlen(vptr)) {
-        if (NULL != cptr) {
-            *cptr = '\0';
-        }
-        /* determine this node's prefix by looking for first non-alpha char */
-        fullval = false;
-        len = strlen(vptr);
-        startnum = -1;
-        memset(prefix, 0, PMIX_MAX_NODE_PREFIX);
-        numdigits = 0;
-        for (i=0, j=0; i < len; i++) {
-            if (!isalpha(vptr[i])) {
-                /* found a non-alpha char */
-                if (!isdigit(vptr[i])) {
-                    /* if it is anything but a digit, we just use
-                     * the entire name
-                     */
-                    fullval = true;
-                    break;
-                }
-                /* count the size of the numeric field - but don't
-                 * add the digits to the prefix
-                 */
-                numdigits++;
-                if (startnum < 0) {
-                    /* okay, this defines end of the prefix */
-                    startnum = i;
-                }
-                continue;
-            }
-            if (startnum < 0) {
-                prefix[j++] = vptr[i];
-            }
-        }
-        if (fullval || startnum < 0) {
-            /* can't compress this name - just add it to the list */
-            vreg = PMIX_NEW(pmix_regex_value_t);
-            vreg->prefix = strdup(vptr);
-            pmix_list_append(&vids, &vreg->super);
-            /* move to the next posn */
-            if (NULL == cptr) {
-                break;
-            }
-            vptr = cptr + 1;
-            continue;
-        }
-        /* convert the digits and get any suffix */
-        vnum = strtol(&vptr[startnum], &sfx, 10);
-        if (NULL != sfx) {
-            suffix = strdup(sfx);
-        } else {
-            suffix = NULL;
-        }
-        /* is this value already on our list? */
-        found = false;
-        PMIX_LIST_FOREACH(vreg, &vids, pmix_regex_value_t) {
-            if (0 < strlen(prefix) && NULL == vreg->prefix) {
-                continue;
-            }
-            if (0 == strlen(prefix) && NULL != vreg->prefix) {
-                continue;
-            }
-            if (0 < strlen(prefix) && NULL != vreg->prefix
-                && 0 != strcmp(prefix, vreg->prefix)) {
-                continue;
-            }
-            if (NULL == suffix && NULL != vreg->suffix) {
-                continue;
-            }
-            if (NULL != suffix && NULL == vreg->suffix) {
-                continue;
-            }
-            if (NULL != suffix && NULL != vreg->suffix &&
-                0 != strcmp(suffix, vreg->suffix)) {
-                continue;
-            }
-            if (numdigits != vreg->num_digits) {
-                continue;
-            }
-            /* found a match - flag it */
-            found = true;
-            /* get the last range on this nodeid - we do this
-             * to preserve order
-             */
-            range = (pmix_regex_range_t*)pmix_list_get_last(&vreg->ranges);
-            if (NULL == range) {
-                /* first range for this value */
-                range = PMIX_NEW(pmix_regex_range_t);
-                range->start = vnum;
-                range->cnt = 1;
-                pmix_list_append(&vreg->ranges, &range->super);
-                break;
-            }
-            /* see if the value is out of sequence */
-            if (vnum != (range->start + range->cnt)) {
-                /* start a new range */
-                range = PMIX_NEW(pmix_regex_range_t);
-                range->start = vnum;
-                range->cnt = 1;
-                pmix_list_append(&vreg->ranges, &range->super);
-                break;
-            }
-            /* everything matches - just increment the cnt */
-            range->cnt++;
-            break;
-        }
-        if (!found) {
-            /* need to add it */
-            vreg = PMIX_NEW(pmix_regex_value_t);
-            if (0 < strlen(prefix)) {
-                vreg->prefix = strdup(prefix);
-            }
-            if (NULL != suffix) {
-                vreg->suffix = strdup(suffix);
-            }
-            vreg->num_digits = numdigits;
-            pmix_list_append(&vids, &vreg->super);
-            /* record the first range for this value - we took
-             * care of values we can't compress above
-             */
-            range = PMIX_NEW(pmix_regex_range_t);
-            range->start = vnum;
-            range->cnt = 1;
-            pmix_list_append(&vreg->ranges, &range->super);
-        }
-        if (NULL != suffix) {
-            free(suffix);
-        }
-        /* move to the next posn */
-        if (NULL == cptr) {
-            break;
-        }
-        vptr = cptr + 1;
-    }
-    free(vsave);
-
-    /* begin constructing the regular expression */
-    while (NULL != (vreg = (pmix_regex_value_t*)pmix_list_remove_first(&vids))) {
-        /* if no ranges, then just add the name */
-        if (0 == pmix_list_get_size(&vreg->ranges)) {
-            if (NULL != vreg->prefix) {
-                pmix_argv_append_nosize(&regexargs, tmp);
-            }
-            PMIX_RELEASE(vreg);
-            continue;
-        }
-        /* start the regex for this value with the prefix */
-        if (NULL != vreg->prefix) {
-            if (0 > asprintf(&tmp, "%s[%d:", vreg->prefix, vreg->num_digits)) {
-                return PMIX_ERR_NOMEM;
-            }
-        } else {
-            if (0 > asprintf(&tmp, "[%d:", vreg->num_digits)) {
-                return PMIX_ERR_NOMEM;
-            }
-        }
-        /* add the ranges */
-        while (NULL != (range = (pmix_regex_range_t*)pmix_list_remove_first(&vreg->ranges))) {
-            if (1 == range->cnt) {
-                if (0 > asprintf(&tmp2, "%s%d,", tmp, range->start)) {
-                    return PMIX_ERR_NOMEM;
-                }
-            } else {
-                if (0 > asprintf(&tmp2, "%s%d-%d,", tmp, range->start, range->start + range->cnt - 1)) {
-                    return PMIX_ERR_NOMEM;
-                }
-            }
-            free(tmp);
-            tmp = tmp2;
-            PMIX_RELEASE(range);
-        }
-        /* replace the final comma */
-        tmp[strlen(tmp)-1] = ']';
-        if (NULL != vreg->suffix) {
-            /* add in the suffix, if provided */
-            if (0 > asprintf(&tmp2, "%s%s", tmp, vreg->suffix)) {
-                return PMIX_ERR_NOMEM;
-            }
-            free(tmp);
-            tmp = tmp2;
-        }
-        pmix_argv_append_nosize(&regexargs, tmp);
-        free(tmp);
-        PMIX_RELEASE(vreg);
-    }
-
-    /* assemble final result */
-    tmp = pmix_argv_join(regexargs, ',');
-    if (0 > asprintf(regexp, "pmix[%s]", tmp)) {
-        return PMIX_ERR_NOMEM;
-    }
-    free(tmp);
-
-    /* cleanup */
-    pmix_argv_free(regexargs);
-
-    PMIX_DESTRUCT(&vids);
-    return PMIX_SUCCESS;
+    return pmix_preg.generate_node_regex(input, regexp);
 }
 
 PMIX_EXPORT pmix_status_t PMIx_generate_ppn(const char *input, char **regexp)
 {
-    char **ppn, **npn;
-    int i, j, start, end;
-    pmix_regex_value_t *vreg;
-    pmix_regex_range_t *rng;
-    pmix_list_t nodes;
-    char *tmp, *tmp2;
-    char *cptr;
-
     PMIX_ACQUIRE_THREAD(&pmix_global_lock);
     if (pmix_globals.init_cntr <= 0) {
         PMIX_RELEASE_THREAD(&pmix_global_lock);
@@ -1456,111 +1229,7 @@ PMIX_EXPORT pmix_status_t PMIx_generate_ppn(const char *input, char **regexp)
     }
     PMIX_RELEASE_THREAD(&pmix_global_lock);
 
-    /* define the default */
-    *regexp = NULL;
-
-    /* setup the list of results */
-    PMIX_CONSTRUCT(&nodes, pmix_list_t);
-
-    /* split the input by node */
-    ppn = pmix_argv_split(input, ';');
-
-    /* for each node, split the input by comma */
-    for (i=0; NULL != ppn[i]; i++) {
-        rng = NULL;
-        /* create a record for this node */
-        vreg = PMIX_NEW(pmix_regex_value_t);
-        pmix_list_append(&nodes, &vreg->super);
-        /* split the input for this node */
-        npn = pmix_argv_split(ppn[i], ',');
-        /* look at each element */
-        for (j=0; NULL != npn[j]; j++) {
-            /* is this a range? */
-            if (NULL != (cptr = strchr(npn[j], '-'))) {
-                /* terminate the string */
-                *cptr = '\0';
-                ++cptr;
-                start = strtol(npn[j], NULL, 10);
-                end = strtol(cptr, NULL, 10);
-                /* are we collecting a range? */
-                if (NULL == rng) {
-                    /* no - better start one */
-                    rng = PMIX_NEW(pmix_regex_range_t);
-                    rng->start = start;
-                    rng->cnt = end - start + 1;
-                    pmix_list_append(&vreg->ranges, &rng->super);
-                } else {
-                    /* is this a continuation of the current range? */
-                    if (start == (rng->start + rng->cnt)) {
-                        /* just add it to the end of this range */
-                        rng->cnt++;
-                    } else {
-                        /* nope, there is a break - create new range */
-                        rng = PMIX_NEW(pmix_regex_range_t);
-                        rng->start = start;
-                        rng->cnt = end - start + 1;
-                        pmix_list_append(&vreg->ranges, &rng->super);
-                    }
-                }
-            } else {
-                /* single rank given */
-                start = strtol(npn[j], NULL, 10);
-                /* are we collecting a range? */
-                if (NULL == rng) {
-                    /* no - better start one */
-                    rng = PMIX_NEW(pmix_regex_range_t);
-                    rng->start = start;
-                    rng->cnt = 1;
-                    pmix_list_append(&vreg->ranges, &rng->super);
-                } else {
-                    /* is this a continuation of the current range? */
-                    if (start == (rng->start + rng->cnt)) {
-                        /* just add it to the end of this range */
-                        rng->cnt++;
-                    } else {
-                        /* nope, there is a break - create new range */
-                        rng = PMIX_NEW(pmix_regex_range_t);
-                        rng->start = start;
-                        rng->cnt = 1;
-                        pmix_list_append(&vreg->ranges, &rng->super);
-                    }
-                }
-            }
-        }
-        pmix_argv_free(npn);
-    }
-    pmix_argv_free(ppn);
-
-
-    /* begin constructing the regular expression */
-    tmp = strdup("pmix[");
-    PMIX_LIST_FOREACH(vreg, &nodes, pmix_regex_value_t) {
-        while (NULL != (rng = (pmix_regex_range_t*)pmix_list_remove_first(&vreg->ranges))) {
-            if (1 == rng->cnt) {
-                if (0 > asprintf(&tmp2, "%s%d,", tmp, rng->start)) {
-                    return PMIX_ERR_NOMEM;
-                }
-            } else {
-                if (0 > asprintf(&tmp2, "%s%d-%d,", tmp, rng->start, rng->start + rng->cnt - 1)) {
-                    return PMIX_ERR_NOMEM;
-                }
-            }
-            free(tmp);
-            tmp = tmp2;
-            PMIX_RELEASE(rng);
-        }
-        /* replace the final comma */
-        tmp[strlen(tmp)-1] = ';';
-    }
-
-    /* replace the final semi-colon */
-    tmp[strlen(tmp)-1] = ']';
-
-    /* assemble final result */
-    *regexp = tmp;
-
-    PMIX_LIST_DESTRUCT(&nodes);
-    return PMIX_SUCCESS;
+    return pmix_preg.generate_ppn(input, regexp);
 }
 
 static void _setup_op(pmix_status_t rc, void *cbdata)
